consolidate constant propagation variations
